requirements/requirements.txt
asgiref==3.3.1
Django==3.1.4
django-ckeditor==6.0.0
django-js-asset==1.2.2
djangorestframework==3.12.2
gunicorn==20.0.4
nginx==0.0.1
Pillow==8.0.1
postgres==3.0.0
psycopg2-binary==2.8.6
psycopg2-pool==1.1
pytz==2020.4
sqlparse==0.4.1
